# backend/contracts/rustrag.openapi.yaml
openapi: 3.1.0
info:
  title: RustRAG API
  version: 0.1.0
  description: |
    Runtime API contract for the current RustRAG backend implementation.

    This contract is hand-maintained from the implemented Axum routes in `backend/src/interfaces/http`.
    It aims to reflect the currently available endpoints, request bodies, response shapes,
    auth requirements, and common error envelopes as accurately as practical.
servers:
  - url: http://localhost:8095
    description: Local default
security:
  - bearerAuth: []
...
paths:
...
  /v1/projects:
    get:
      tags: [projects]
      summary: List projects
      operationId: listProjects
      security: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSummary'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [projects]
      summary: Create project
      operationId: createProject
      description: Requires `projects:write`, `workspace:admin`, or an `instance_admin` token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '200':
          description: Created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/provider-accounts:
    get:
      tags: [providers]
      summary: List provider accounts
      operationId: listProviderAccounts
      security: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Provider account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProviderAccountSummary'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [providers]
      summary: Create provider account
      operationId: createProviderAccount
      description: Requires `providers:admin`, `workspace:admin`, or an `instance_admin` token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProviderAccountRequest'
      responses:
        '200':
          description: Created provider account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderAccountSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/model-profiles:
    get:
      tags: [providers]
      summary: List model profiles
      operationId: listModelProfiles
      security: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Model profile list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelProfileSummary'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [providers]
      summary: Create model profile
      operationId: createModelProfile
      description: Requires `providers:admin`, `workspace:admin`, or an `instance_admin` token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelProfileRequest'
      responses:
        '200':
          description: Created model profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProfileSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/sources:
    get:
      tags: [ingestion]
      summary: List sources
      operationId: listSources
      security: []
      parameters:
        - $ref: '#/components/parameters/projectIdOptional'
      responses:
        '200':
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceSummary'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [ingestion]
      summary: Create source
      operationId: createSource
      description: Requires `documents:write`, `workspace:admin`, or an `instance_admin` token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        '200':
          description: Created source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/ingestion-jobs:
    get:
      tags: [ingestion]
      summary: List ingestion jobs
      operationId: listIngestionJobs
      security: []
      parameters:
        - $ref: '#/components/parameters/projectIdOptional'
      responses:
        '200':
          description: Ingestion job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IngestionJobSummary'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [ingestion]
      summary: Create ingestion job
      operationId: createIngestionJob
      description: Requires `documents:write`, `workspace:admin`, or an `instance_admin` token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIngestionJobRequest'
      responses:
        '200':
          description: Created ingestion job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token minted by `POST /v1/auth/tokens`.
        Note: some read/list endpoints are currently public and do not require auth.
  parameters:
    workspaceId:
      in: query
      name: workspace_id
      required: false
      schema:
        type: string
        format: uuid
    projectIdOptional:
      in: query
      name: project_id
      required: false
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
          examples:
            apiErrorEnvelope:
              value:
                error: unauthorized
        text/plain:
          schema:
            type: string
          examples:
            authExtractorError:
              value: missing authorization header
      headers:
        WWW-Authenticate:
          required: false
          schema:
            type: string
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
        text/plain:
          schema:
            type: string
  schemas:
    ApiErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProjectSummary:
      type: object
      required: [id, workspace_id, slug, name]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: [string, 'null']
    CreateProjectRequest:
      type: object
      required: [workspace_id, slug, name]
      properties:
        workspace_id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: [string, 'null']
    ProviderAccountSummary:
      type: object
      required: [id, workspace_id, provider_kind, label, status]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        provider_kind:
          type: string
          example: openai
        label:
          type: string
        status:
          type: string
          example: active
    CreateProviderAccountRequest:
      type: object
      required: [workspace_id, provider_kind, label]
      properties:
        workspace_id:
          type: string
          format: uuid
        provider_kind:
          type: string
        label:
          type: string
        api_base_url:
          type: [string, 'null']
          format: uri
    ModelProfileSummary:
      type: object
      required: [id, workspace_id, provider_account_id, profile_kind, model_name]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        provider_account_id:
          type: string
          format: uuid
        profile_kind:
          type: string
          example: chat
        model_name:
          type: string
          example: gpt-4.1-mini
    CreateModelProfileRequest:
      type: object
      required: [workspace_id, provider_account_id, profile_kind, model_name]
      properties:
        workspace_id:
          type: string
          format: uuid
        provider_account_id:
          type: string
          format: uuid
        profile_kind:
          type: string
        model_name:
          type: string
        temperature:
          type: [number, 'null']
        max_output_tokens:
          type: [integer, 'null']
          format: int32
    SourceSummary:
      type: object
      required: [id, project_id, source_kind, label, status]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        source_kind:
          type: string
          example: upload
        label:
          type: string
        status:
          type: string
          example: active
    CreateSourceRequest:
      type: object
      required: [project_id, source_kind, label]
      properties:
        project_id:
          type: string
          format: uuid
        source_kind:
          type: string
        label:
          type: string
    IngestionJobSummary:
      type: object
      required: [id, project_id, trigger_kind, status, stage]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        source_id:
          type: [string, 'null']
          format: uuid
        trigger_kind:
          type: string
        status:
          type: string
          example: queued
        stage:
          type: string
          example: created
    CreateIngestionJobRequest:
      type: object
      required: [project_id, trigger_kind]
      properties:
        project_id:
          type: string
          format: uuid
        source_id:
          type: [string, 'null']
          format: uuid
        trigger_kind:
          type: string
        requested_by:
          type: [string, 'null']
...